% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{alexa}
\alias{alexa}
\title{Alexa Top 1 Million Sites}
\format{A data frame with 1000024 rows and 2 variables}
\usage{
data(alexa)
}
\description{
Alexa Top 1 Million Sites
\itemize{
  \item \code{rank}. rank of the domain (int)
  \item \code{domain}. the domain name (chr)
}
}
\note{
Last updated 2014-08-09
}
\seealso{
\itemize{
  \item IANA - \url{http://aws.amazon.com/alexa-top-sites/}
}
}
\keyword{datasets}

