\name{asninfo}
\alias{asninfo}
\title{Return ASN info from Team CYNRU DNS lookup service}
\usage{
asninfo(asn = "AS23028")
}
\arguments{
  \item{ASN}{number (with or without "AS" prefixed) to
  lookup (character vector)}
}
\value{
list containing named ASN attributes
}
\description{
Return ASN info from Team CYNRU DNS lookup service
}

