% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CNAME}
\alias{CNAME}
\title{Vectorized version of \link{resolv_cname}}
\usage{
CNAME(fqdn, nameserver = NA_character_, showWarnings = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}
}
\value{
list
}
\description{
Vectorized version of \link{resolv_cname}
}

