% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{dnscalc}
\alias{dnscalc}
\title{Return simple math computations via DNS}
\usage{
dnscalc(a = 0:99, b = 0:99, op = c("add", "sub", "mul", "div"))
}
\arguments{
\item{a}{first number (0-99)}

\item{b}{second number (0-99)}

\item{op}{calculation to perform (add|sub|mul|div)}
}
\value{
result (in IP address octet notation)
}
\description{
Return simple math computations via DNS
}
\seealso{
http://www.isi.edu/touch/tools/dns-calc.html
}

