% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MX}
\alias{MX}
\title{Vectorized version of \link{resolv_mx}}
\usage{
MX(domain, nameserver = NA_character_, showWarnings = FALSE, full = FALSE)
}
\arguments{
\item{domain}{input character vector (domain name)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}

\item{full}{include full record response information in results (bool)}
}
\value{
named vector or list
}
\description{
Vectorized version of \link{resolv_mx}
}

