% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{NS}
\alias{NS}
\title{Vectorized version of \link{resolv_ns}}
\usage{
NS(fqdn, nameserver = NA_character_, showWarnings = FALSE, full = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}

\item{full}{include full record response information in results (bool)}
}
\value{
list
}
\description{
Vectorized version of \link{resolv_ns}
}

