% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{resolv_ns}
\alias{resolv_ns}
\title{Returns the DNS NS records for a given FQDN}
\usage{
resolv_ns(fqdn, nameserver = NA_character_, showWarnings = FALSE,
  full = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}

\item{full}{include full record response information in results (bool)}
}
\value{
vector or data frame (if \code{full}==\code{TRUE}) of NS records or \code{character(0)} if none
}
\description{
Returns the DNS NS records for a given FQDN
}
\examples{
require(resolv)

resolv_ns("www.paypal.com")
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/}
}

