% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{resolv_srv}
\alias{resolv_srv}
\title{Returns the DNS SRV records for a given FQDN}
\usage{
resolv_srv(fqdn, nameserver = NA_character_, showWarnings = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}
}
\value{
list of SRV records (named fields) or an empty list if none
}
\description{
Returns the DNS SRV records for a given FQDN
}
\examples{
require(resolv)
library(plyr)

## google talk provides a good example for this
ldply(resolv_srv("_xmpp-server._tcp.gmail.com."), unlist)
 priority weight port                         target
1        5      0 5269      xmpp-server.l.google.com.
2       20      0 5269 alt1.xmpp-server.l.google.com.
3       20      0 5269 alt2.xmpp-server.l.google.com.
4       20      0 5269 alt3.xmpp-server.l.google.com.
5       20      0 5269 alt4.xmpp-server.l.google.com.
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/}
}

