% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{TXT}
\alias{TXT}
\title{Vectorized version of \link{resolv_txt}}
\usage{
TXT(fqdn, nameserver = NA_character_, showWarnings = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}
}
\value{
named vector or list
}
\description{
Vectorized version of \link{resolv_txt}
}

