\name{asninfo}
\alias{asninfo}
\title{Return ASN info from Team CYNRU DNS lookup service}
\usage{
asninfo(asn = "AS23028")
}
\arguments{
  \item{ASN}{number (with or without "AS" prefixed) to
  lookup (character vector)}
}
\value{
list containing named ASN attributes
}
\description{
Return ASN info from Team CYNRU DNS lookup service
}
\seealso{
Other resolv: \code{\link{dnscalc}}; \code{\link{ip2asn}};
\code{\link{resolv_a}}; \code{\link{resolv_cname}};
\code{\link{resolv_mx}}; \code{\link{resolv_ptr}};
\code{\link{resolv_srv}}; \code{\link{resolv_txt}};
\code{\link{wikidns}}
}

