% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ip2asn}
\alias{ip2asn}
\title{Return ASN info from Team CYNRU DNS lookup service}
\usage{
ip2asn(ip = "216.90.108.31")
}
\arguments{
\item{ip}{address to lookup (character vector)}
}
\value{
data frame containing named ASN attributes
}
\description{
Pretty much provided as an example of data services. You
should use the one in the \link{iptools} or \link{netintel} packages
if you are serious about doing IP/ASN lookups.
}

