\name{ip2asn}
\alias{ip2asn}
\title{Return ASN info from Team CYNRU DNS lookup service}
\usage{
ip2asn(ip = "216.90.108.31")
}
\arguments{
  \item{ip}{address to lookup (character vector)}
}
\value{
list containing named ASN attributes
}
\description{
Return ASN info from Team CYNRU DNS lookup service
}
\seealso{
Other resolv: \code{\link{asninfo}}; \code{\link{dnscalc}};
\code{\link{resolv_a}}; \code{\link{resolv_cname}};
\code{\link{resolv_mx}}; \code{\link{resolv_ptr}};
\code{\link{resolv_srv}}; \code{\link{resolv_txt}};
\code{\link{wikidns}}
}

