% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{ip2asn}
\alias{ip2asn}
\title{Return ASN info from Team CYNRU DNS lookup service}
\usage{
ip2asn(ip = "216.90.108.31")
}
\arguments{
\item{ip}{address to lookup (character vector)}
}
\value{
list containing named ASN attributes
}
\description{
Return ASN info from Team CYNRU DNS lookup service
}

