% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{resolv-package}
\alias{resolv}
\alias{resolv-package}
\title{Wrapper to ldns library for DNS calls from R}
\description{
Wrapper to ldns library for DNS calls from R
}
\examples{
\dontrun{
require(resolv)
library(plyr)

## google talk provides a good example for this
ldply(resolv_srv("_xmpp-server._tcp.gmail.com."), unlist)
## priority weight port                         target
## 1        5      0 5269      xmpp-server.l.google.com.
## 2       20      0 5269 alt1.xmpp-server.l.google.com.
## 3       20      0 5269 alt2.xmpp-server.l.google.com.
## 4       20      0 5269 alt3.xmpp-server.l.google.com.
## 5       20      0 5269 alt4.xmpp-server.l.google.com.

where www.nasa.gov hosts
resolv_a("www.nasa.gov")
## [1] "69.28.187.45"    "208.111.161.110"

resolv_ptr("69.28.187.45")
## [1] "cds355.iad.llnw.net."

# seekrit URLs
browseURL(gsub("\\"", "", resolv_txt("google-public-dns-a.google.com")))

}
}
\author{
Bob Rudis <bob@rudis.net>
}
\references{
\url{http://www.nlnetlabs.nl/projects/ldns/}
}
\seealso{
\itemize{
  \item \url{https://github.com/hrbrmstr/resolv}
  \item \url{http://www.nlnetlabs.nl/projects/ldns/}
  \item \url{http://dds.ec/blog/posts/2014/Apr/making-better-dns-txt-record-lookups-with-rcpp/}
}
}

