% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{resolv-package}
\alias{resolv}
\alias{resolv-package}
\title{Wrapper to ldns library for DNS calls from R}
\description{
Wrapper to ldns library for DNS calls from R
}
\examples{
\dontrun{
require(resolv)
library(plyr)

## google talk provides a good example for this
resolv_srv("_xmpp-server._tcp.gmail.com.")

##                           fqdn priority weight port                         target
## 1 _xmpp-server._tcp.gmail.com.        5      0 5269      xmpp-server.l.google.com.
## 2 _xmpp-server._tcp.gmail.com.       20      0 5269 alt1.xmpp-server.l.google.com.
## 3 _xmpp-server._tcp.gmail.com.       20      0 5269 alt2.xmpp-server.l.google.com.
## 4 _xmpp-server._tcp.gmail.com.       20      0 5269 alt3.xmpp-server.l.google.com.
## 5 _xmpp-server._tcp.gmail.com.       20      0 5269 alt4.xmpp-server.l.google.com.

## where www.nasa.gov hosts
resolv_a("www.nasa.gov", full=TRUE)

##           fqdn       address                owner class ttl
## 1 www.nasa.gov 69.28.157.198 iznasa.hs.llnwd.net.     1 274

resolv_ptr("69.28.187.45")
## [1] "cds355.iad.llnw.net."

## vectorized
A(c("dds.ec", "rud.is"), full=TRUE)

## $dds.ec
##     fqdn       address   owner class ttl
## 1 dds.ec 162.243.111.4 dds.ec.     1  87
##
## $rud.is
##     fqdn        address   owner class   ttl
## 1 rud.is 184.106.97.102 rud.is.     1 17781
# seekrit URLs
browseURL(gsub("\\"", "", resolv_txt("google-public-dns-a.google.com")))

}
}
\author{
Bob Rudis <bob@rudis.net>
}
\references{
\url{http://www.nlnetlabs.nl/projects/ldns/}
}
\seealso{
\itemize{
  \item \url{https://github.com/hrbrmstr/resolv}
  \item \url{http://www.nlnetlabs.nl/projects/ldns/}
  \item \url{http://dds.ec/blog/posts/2014/Apr/making-better-dns-txt-record-lookups-with-rcpp/}
}
}

