% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{resolv_cname}
\alias{resolv_cname}
\title{Returns the DNS CNAME records for a given FQDN}
\usage{
resolv_cname(fqdn, nameserver = NA_character_, showWarnings = FALSE)
}
\arguments{
\item{fqdn}{input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}
}
\value{
vector of CNAME records or \code{character(0)} if none
}
\description{
Returns the DNS CNAME records for a given FQDN
}
\examples{
require(resolv)

resolv_cname("www.paypal.com")
[1] "www.paypal.com.akadns.net."
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/}
}

