% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolv_mx}
\alias{resolv_mx}
\title{Returns the DNS MX records for a given domain}
\usage{
resolv_mx(fqdn, nameserver = NA_character_, showWarnings = FALSE,
  full = FALSE)
}
\arguments{
\item{fqdn}{input character vector (domain name)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}

\item{full}{include full record response information in results (bool)}
}
\value{
data frame of MX records (preference & exchange; +owner,class,ttl if \code{full}==\code{TRUE}) or an empty data frame if none
}
\description{
Returns the DNS MX records for a given domain
}
\examples{
\dontrun{
require(resolv)

resolv_mx("rudis.net", full=TRUE)
##        fqdn prefernece                 exchange      owner class ttl
## 1 rudis.net          1      aspmx.l.google.com. rudis.net.     1 599
## 2 rudis.net          5 alt1.aspmx.l.google.com. rudis.net.     1 599
## 3 rudis.net          5 alt2.aspmx.l.google.com. rudis.net.     1 599
## 4 rudis.net         10   aspmx2.googlemail.com. rudis.net.     1 599
## 5 rudis.net         10   aspmx3.googlemail.com. rudis.net.     1 599
## 6 rudis.net         10   aspmx4.googlemail.com. rudis.net.     1 599
## 7 rudis.net         10   aspmx5.googlemail.com. rudis.net.     1 599
## 8 rudis.net        100  mx-caprica.easydns.com. rudis.net.     1 599
}
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/} (cool DNS MX hacks vla \url{https://twitter.com/habbie/status/460067198586081280})
}

