% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolv_mx}
\alias{resolv_mx}
\title{Returns the DNS MX records for a given domain}
\usage{
resolv_mx(domain, nameserver = NA_character_, showWarnings = FALSE)
}
\arguments{
\item{domain}{input character vector (domain name)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}
}
\value{
list of MX records (preference & exchange) or \code{NULL} if none
}
\description{
Returns the DNS MX records for a given domain
}
\examples{
require(resolv)

unlist(resolv_mx("securitymetrics.org"))
           preference               exchange
                  "0" "securitymetrics.org."

## get the MX records for Google
unlist(sapply(resolv_mx("rud.is"), "[", "exchange"), use.names=FALSE)
[1] "aspmx.l.google.com."      "alt1.aspmx.l.google.com."
[3] "alt2.aspmx.l.google.com." "aspmx2.googlemail.com."
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/} (cool DNS MX hacks vla \url{https://twitter.com/habbie/status/460067198586081280})
}

