% Generated by roxygen2 (4.0.1): do not edit by hand
\name{resolv_ptr}
\alias{resolv_ptr}
\title{Returns the DNS PTR records for a given IP address}
\usage{
resolv_ptr(ip, nameserver = NA_character_, showWarnings = FALSE,
  full = FALSE)
}
\arguments{
\item{ip}{address input character vector (FQDN)}

\item{nameserver}{the nameserver to send the request to (optional; uses standard resolver behavior if not specified)}

\item{showWarnings}{display R warning messages (bool)}

\item{full}{include full record response information in results (bool)}
}
\value{
vector or data frame (if \code{full}==\code{TRUE}) of PTR records or \code{character(0)} if none
}
\description{
Returns the DNS PTR records for a given IP address
}
\examples{
\dontrun{
require(resolv)

# where www.nasa.gov hosts
resolv_a("www.nasa.gov")
[1] "69.28.187.45"    "208.111.161.110"

resolv_ptr("69.28.187.45")
[1] "cds355.iad.llnw.net."

## big one - truncated output
resolv_ptr("17.149.160.49")
  [1] "asto.re."                   "next.com."
  [3] "qtml.com."                  "qttv.net."
  [5] "apple.com."                 "apple.info."
  [7] "ikids.com."                 "qt-tv.net."
  [9] "carbon.com."                "eworld.com."
...
[131] "quicktimestreaming.net."    "publishing-research.com."
[133] "publishing-research.org."   "applefinalcutproworld.com."
[135] "applefinalcutproworld.net." "applefinalcutproworld.org."
}
}
\seealso{
\url{http://www.nlnetlabs.nl/projects/ldns/}

\url{http://www.cambus.net/interesting-dns-hacks/}
}

