% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{wikidns}
\alias{wikidns}
\title{Return wikipedia topic via DNS}
\usage{
wikidns(term = "bayes")
}
\arguments{
\item{term}{wikipedia term to lookup}
}
\value{
vector of TXT record results for term description
}
\description{
Return wikipedia topic via DNS
}
\seealso{
https://dgl.cx/wikipedia-dns
}

